Practice table creation
In this code, I created a store inventory table with 5 columns of information. I then inserted 15 potential items that would be sold in the store along with their item id, name, quantity in inventory, price, and number sold. Finally I used the SELECT statement to order by inventory by price and used aggregate functions to understand the amount of items that have been sold. All of these numbers are from a fictional store.