JOINS
In this code, more advanced SQL techniques are learned, practiced, and applied to a dataset. This displays Inner, outer, left, and right joins.  Using matching columns to connect data that aren't directly related by a primary or foreign key is explored and various relationships are made between tables in the data sets. CROSS JOIN is also used on its own and in conjunction with the WHERE clause.  UNION and UNION ALL are also touched on at the end. This was all done in MySQL.